+++ aldrin/CHANGELOG.md
@@ -29,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `DiscovererEvent` as a variant of `DiscovererEventRef` that doesn't borrow the discoverer.
 - Add `DiscovererStream`.
 - Add `Reply::abort()` and `low_level::Reply::abort()`.
+- Add `Promise::is_aborted()`, `Promise::poll_aborted()`, `Promise::aborted()`,
+  `low_level::Promise::is_aborted()`, `low_level::Promise::poll_aborted()` and
+  `low_level::Promise::aborted()`;
 
 ### Fixed